# docs/openapi.json
{
  "components" : {
    "responses" : {
      "BadRequest" : {
        "content" : {
          "application/problem+json" : {
            "examples" : {
              "validation-error" : {
                "summary" : "Invalid request",
                "value" : {
                  "errors" : [ {
                    "detail" : "must not be blank",
                    "pointer" : "#/title"
                  } ],
                  "status" : 400,
                  "title" : "Invalid request",
                  "type" : "https://github.com/jubasse/spring-boot-tasks-api/blob/main/docs/problems.md#validation-error"
                }
              }
            },
            "schema" : {
              "$ref" : "#/components/schemas/ValidationProblem"
            }
          }
        },
        "description" : "Invalid request: see errors for each invalid value."
      },
      "ContentTooLarge" : {
        "content" : {
          "application/problem+json" : {
            "schema" : {
              "$ref" : "#/components/schemas/Problem"
            }
          }
        },
        "description" : "A file is larger than its limit."
      },
      "Forbidden" : {
        "description" : "The account may not do this. Task operations need an active account (verified email, not disabled); other rules are in the operation's description.",
        "headers" : {
          "WWW-Authenticate" : {
            "description" : "Bearer error=\"insufficient_scope\".",
            "schema" : {
              "type" : "string"
            }
          }
        }
      },
...
      "NotFound" : {
        "content" : {
          "application/problem+json" : {
            "schema" : {
              "$ref" : "#/components/schemas/Problem"
            }
          }
        },
        "description" : "The resource does not exist."
      },
...
      "Unauthorized" : {
        "description" : "Missing, expired or invalid access token.",
        "headers" : {
          "WWW-Authenticate" : {
            "description" : "Bearer, with the reason of the refusal.",
            "schema" : {
              "type" : "string"
            }
          }
        }
      },
      "UnexpectedError" : {
        "content" : {
          "application/problem+json" : {
            "schema" : {
              "$ref" : "#/components/schemas/Problem"
            }
          }
        },
        "description" : "An unexpected failure on the server."
      },
      "UnsupportedMediaType" : {
        "content" : {
          "application/problem+json" : {
            "schema" : {
              "$ref" : "#/components/schemas/Problem"
            }
          }
        },
        "description" : "The file type is not accepted, whatever its name or declared type."
      }
    },
    "schemas" : {
...
      "CreateTaskDto" : {
        "properties" : {
          "assignedTo" : {
            "format" : "uuid",
            "type" : "string"
          },
          "description" : {
            "type" : "string"
          },
          "dueAt" : {
            "format" : "date-time",
            "type" : "string"
          },
          "priority" : {
            "enum" : [ "LOW", "MEDIUM", "HIGH", "URGENT" ],
            "type" : "string"
          },
          "reference" : {
            "maxLength" : 30,
            "minLength" : 0,
            "type" : "string"
          },
          "title" : {
            "maxLength" : 255,
            "minLength" : 0,
            "type" : "string"
          }
        },
        "required" : [ "reference", "title" ],
        "type" : "object"
      },
...
      "PagedModelTaskCommentResponseDto" : {
        "properties" : {
          "content" : {
            "items" : {
              "$ref" : "#/components/schemas/TaskCommentResponseDto"
            },
            "type" : "array"
          },
          "page" : {
            "$ref" : "#/components/schemas/PageMetadata"
          }
        },
        "required" : [ "content", "page" ],
        "type" : "object"
      },
...
      "PagedModelTaskResponseDto" : {
        "properties" : {
          "content" : {
            "items" : {
              "$ref" : "#/components/schemas/TaskResponseDto"
            },
            "type" : "array"
          },
          "page" : {
            "$ref" : "#/components/schemas/PageMetadata"
          }
        },
        "required" : [ "content", "page" ],
        "type" : "object"
      },
...
      "Problem" : {
        "description" : "A problem document (RFC 9457). docs/problems.md lists every problem and what to do.",
        "properties" : {
          "detail" : {
            "description" : "This occurrence, for a person to read. It follows Accept-Language: do not parse it.",
            "type" : "string"
          },
          "instance" : {
            "description" : "The path of the request.",
            "format" : "uri-reference",
            "type" : "string"
          },
          "status" : {
            "format" : "int32",
            "maximum" : 599,
            "minimum" : 100,
            "type" : "integer"
          },
          "title" : {
            "description" : "Short summary, the same for every occurrence: the problem type's title, or the status phrase.",
            "type" : "string"
          },
          "type" : {
            "description" : "What went wrong; absent when the status says it all (about:blank). It links to the documentation of the problem.",
            "format" : "uri",
            "type" : "string"
          }
        },
        "required" : [ "status", "title" ],
        "type" : "object"
      },
...
      "TaskCommentDto" : {
        "properties" : {
          "body" : {
            "maxLength" : 10000,
            "minLength" : 0,
            "type" : "string"
          }
        },
        "required" : [ "body" ],
        "type" : "object"
      },
...
      "TaskResponseDto" : {
        "properties" : {
          "archivedAt" : {
            "format" : "date-time",
            "type" : "string"
          },
          "assignedTo" : {
            "$ref" : "#/components/schemas/UserProfileResponseDto"
          },
          "blockedAt" : {
            "format" : "date-time",
            "type" : "string"
          },
          "blockedReason" : {
            "type" : "string"
          },
          "cancelledAt" : {
            "format" : "date-time",
            "type" : "string"
          },
          "cancelledReason" : {
            "type" : "string"
          },
          "completedAt" : {
            "format" : "date-time",
            "type" : "string"
          },
          "createdBy" : {
            "$ref" : "#/components/schemas/UserProfileResponseDto"
          },
          "description" : {
            "type" : "string"
          },
          "dueAt" : {
            "format" : "date-time",
            "type" : "string"
          },
          "id" : {
            "format" : "uuid",
            "type" : "string"
          },
          "priority" : {
            "enum" : [ "LOW", "MEDIUM", "HIGH", "URGENT" ],
            "type" : "string"
          },
          "reference" : {
            "type" : "string"
          },
          "status" : {
            "enum" : [ "TO_DO", "IN_PROGRESS", "IN_REVIEW", "BLOCKED", "DONE", "ARCHIVED", "CANCELLED" ],
            "type" : "string"
          },
          "title" : {
            "type" : "string"
          },
          "updatedAt" : {
            "format" : "date-time",
            "type" : "string"
          },
          "version" : {
            "format" : "int64",
            "type" : "integer"
          }
        },
        "type" : "object"
      },
      "UpdateTaskDto" : {
        "properties" : {
          "description" : {
            "type" : "string"
          },
          "dueAt" : {
            "format" : "date-time",
            "type" : "string"
          },
          "priority" : {
            "enum" : [ "LOW", "MEDIUM", "HIGH", "URGENT" ],
            "type" : "string"
          },
          "title" : {
            "maxLength" : 255,
            "minLength" : 0,
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
    },
    "securitySchemes" : {
      "bearer" : {
        "bearerFormat" : "JWT",
        "description" : "The access token returned by POST /api/v1/auth/login or /refresh.",
        "scheme" : "bearer",
        "type" : "http"
      }
    }
  },
...
  "info" : {
    "contact" : {
      "name" : "Tasks API on GitHub",
      "url" : "https://github.com/jubasse/spring-boot-tasks-api"
    },
    "description" : "Tasks, with their comments, files and history, for teams whose members sign up themselves.\n\n**Authentication.** Sign in with `POST /api/v1/auth/login`, then send the access token in an `Authorization: Bearer` header. It lasts 15 minutes: `POST /api/v1/auth/refresh` trades the refresh token for a new pair. Task operations also need an active account: a verified email, not disabled.\n\n**Errors.** Every error is a problem document (RFC 9457, `application/problem+json`), except the 401 and 403 of an access check, which carry only a `WWW-Authenticate` header. When `type` is present, it links to the documentation of that problem.\n\n**Limits.** Login, sign-up, verification email resend and password reset are rate limited: a 429 gives the seconds to wait in `Retry-After`. A page holds at most 100 items.\n",
    "title" : "Tasks API",
    "version" : "development"
  },
  "openapi" : "3.1.0",
  "paths" : {
...
    "/api/v1/tasks" : {
      "get" : {
        "operationId" : "listTasks",
        "parameters" : [ {
          "in" : "query",
          "name" : "status",
          "required" : false,
          "schema" : {
            "enum" : [ "TO_DO", "IN_PROGRESS", "IN_REVIEW", "BLOCKED", "DONE", "ARCHIVED", "CANCELLED" ],
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "assigneeId",
          "required" : false,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "archived",
          "required" : false,
          "schema" : {
            "default" : false,
            "type" : "boolean"
          }
        }, {
          "description" : "Zero-based page index (0..N)",
          "in" : "query",
          "name" : "page",
          "required" : false,
          "schema" : {
            "default" : 0,
            "minimum" : 0,
            "type" : "integer"
          }
        }, {
          "description" : "The size of the page to be returned",
          "in" : "query",
          "name" : "size",
          "required" : false,
          "schema" : {
            "default" : 10,
            "maximum" : 100,
            "minimum" : 1,
            "type" : "integer"
          }
        }, {
          "description" : "Sorting criteria in the format: property,(asc|desc). Default sort order is ascending. Multiple sort criteria are supported.",
          "in" : "query",
          "name" : "sort",
          "required" : false,
          "schema" : {
            "default" : [ "createdAt,DESC" ],
            "items" : {
              "type" : "string"
            },
            "type" : "array"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PagedModelTaskResponseDto"
                }
              }
            },
            "description" : "OK"
          },
          "400" : {
            "$ref" : "#/components/responses/BadRequest"
          },
          "401" : {
            "$ref" : "#/components/responses/Unauthorized"
          },
          "403" : {
            "$ref" : "#/components/responses/Forbidden"
          },
          "default" : {
            "$ref" : "#/components/responses/UnexpectedError"
          }
        },
        "summary" : "List tasks, newest first by default",
        "tags" : [ "Tasks" ]
      },
      "post" : {
        "description" : "**Access:** Any active user; assigning the task on creation requires the ADMIN role.",
        "operationId" : "createTask",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/CreateTaskDto"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/TaskResponseDto"
                }
              }
            },
            "description" : "Created",
            "headers" : {
              "Location" : {
                "description" : "URL of the created resource.",
                "schema" : {
                  "format" : "uri-reference",
                  "type" : "string"
                }
              }
            }
          },
          "400" : {
            "$ref" : "#/components/responses/BadRequest"
          },
          "401" : {
            "$ref" : "#/components/responses/Unauthorized"
          },
          "403" : {
            "$ref" : "#/components/responses/Forbidden"
          },
          "404" : {
            "$ref" : "#/components/responses/NotFound"
          },
          "409" : {
            "content" : {
              "application/problem+json" : {
                "examples" : {
                  "reference-taken" : {
                    "summary" : "Task reference already in use",
                    "value" : {
                      "status" : 409,
                      "title" : "Task reference already in use",
                      "type" : "https://github.com/jubasse/spring-boot-tasks-api/blob/main/docs/problems.md#reference-taken"
                    }
                  }
                },
                "schema" : {
                  "$ref" : "#/components/schemas/Problem"
                }
              }
            },
            "description" : "Conflict: Task reference already in use."
          },
          "422" : {
            "content" : {
              "application/problem+json" : {
                "examples" : {
                  "assignee-not-active" : {
                    "summary" : "Assignee not active",
                    "value" : {
                      "status" : 422,
                      "title" : "Assignee not active",
                      "type" : "https://github.com/jubasse/spring-boot-tasks-api/blob/main/docs/problems.md#assignee-not-active"
                    }
                  }
                },
                "schema" : {
                  "$ref" : "#/components/schemas/Problem"
                }
              }
            },
            "description" : "Unprocessable Content: Assignee not active."
          },
          "default" : {
            "$ref" : "#/components/responses/UnexpectedError"
          }
        },
        "summary" : "Create a task",
        "tags" : [ "Tasks" ]
      }
    },
    "/api/v1/tasks/{id}" : {
      "delete" : {
        "description" : "**Access:** Requires the ADMIN role.",
        "operationId" : "deleteTask",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        } ],
        "responses" : {
          "204" : {
            "description" : "No Content"
          },
          "400" : {
            "$ref" : "#/components/responses/BadRequest"
          },
          "401" : {
            "$ref" : "#/components/responses/Unauthorized"
          },
          "403" : {
            "$ref" : "#/components/responses/Forbidden"
          },
          "404" : {
            "$ref" : "#/components/responses/NotFound"
          },
          "409" : {
            "content" : {
              "application/problem+json" : {
                "examples" : {
                  "version-conflict" : {
                    "summary" : "Changed by another request",
                    "value" : {
                      "status" : 409,
                      "title" : "Changed by another request",
                      "type" : "https://github.com/jubasse/spring-boot-tasks-api/blob/main/docs/problems.md#version-conflict"
                    }
                  }
                },
                "schema" : {
                  "$ref" : "#/components/schemas/Problem"
                }
              }
            },
            "description" : "Conflict: Changed by another request."
          },
          "default" : {
            "$ref" : "#/components/responses/UnexpectedError"
          }
        },
        "summary" : "Delete a task",
        "tags" : [ "Tasks" ]
      },
      "get" : {
        "operationId" : "getTask",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/TaskResponseDto"
                }
              }
            },
            "description" : "OK"
          },
          "400" : {
            "$ref" : "#/components/responses/BadRequest"
          },
          "401" : {
            "$ref" : "#/components/responses/Unauthorized"
          },
          "403" : {
            "$ref" : "#/components/responses/Forbidden"
          },
          "404" : {
            "$ref" : "#/components/responses/NotFound"
          },
          "default" : {
            "$ref" : "#/components/responses/UnexpectedError"
          }
        },
        "summary" : "Get a task",
        "tags" : [ "Tasks" ]
      },
      "patch" : {
        "description" : "**Access:** The task's assignee, or the ADMIN role.",
        "operationId" : "updateTask",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/UpdateTaskDto"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/TaskResponseDto"
                }
              }
            },
            "description" : "OK"
          },
          "400" : {
            "$ref" : "#/components/responses/BadRequest"
          },
          "401" : {
            "$ref" : "#/components/responses/Unauthorized"
          },
          "403" : {
            "$ref" : "#/components/responses/Forbidden"
          },
          "404" : {
            "$ref" : "#/components/responses/NotFound"
          },
          "409" : {
            "content" : {
              "application/problem+json" : {
                "examples" : {
                  "version-conflict" : {
                    "summary" : "Changed by another request",
                    "value" : {
                      "status" : 409,
                      "title" : "Changed by another request",
                      "type" : "https://github.com/jubasse/spring-boot-tasks-api/blob/main/docs/problems.md#version-conflict"
                    }
                  }
                },
                "schema" : {
                  "$ref" : "#/components/schemas/Problem"
                }
              }
            },
            "description" : "Conflict: Changed by another request."
          },
          "default" : {
            "$ref" : "#/components/responses/UnexpectedError"
          }
        },
        "summary" : "Update a task",
        "tags" : [ "Tasks" ]
      }
    },
...
    "/api/v1/tasks/{id}/comments" : {
      "get" : {
        "operationId" : "listComments",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        }, {
          "description" : "Zero-based page index (0..N)",
          "in" : "query",
          "name" : "page",
          "required" : false,
          "schema" : {
            "default" : 0,
            "minimum" : 0,
            "type" : "integer"
          }
        }, {
          "description" : "The size of the page to be returned",
          "in" : "query",
          "name" : "size",
          "required" : false,
          "schema" : {
            "default" : 10,
            "maximum" : 100,
            "minimum" : 1,
            "type" : "integer"
          }
        }, {
          "description" : "Sorting criteria in the format: property,(asc|desc). Default sort order is ascending. Multiple sort criteria are supported.",
          "in" : "query",
          "name" : "sort",
          "required" : false,
          "schema" : {
            "default" : [ "createdAt,ASC" ],
            "items" : {
              "type" : "string"
            },
            "type" : "array"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PagedModelTaskCommentResponseDto"
                }
              }
            },
            "description" : "OK"
          },
          "400" : {
            "$ref" : "#/components/responses/BadRequest"
          },
          "401" : {
            "$ref" : "#/components/responses/Unauthorized"
          },
          "403" : {
            "$ref" : "#/components/responses/Forbidden"
          },
          "404" : {
            "$ref" : "#/components/responses/NotFound"
          },
          "default" : {
            "$ref" : "#/components/responses/UnexpectedError"
          }
        },
        "summary" : "List the comments of a task, oldest first",
        "tags" : [ "Comments" ]
      },
      "post" : {
        "operationId" : "addComment",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/TaskCommentDto"
              }
            },
            "multipart/form-data" : {
              "schema" : {
                "allOf" : [ {
                  "$ref" : "#/components/schemas/TaskCommentDto"
                }, {
                  "properties" : {
                    "files" : {
                      "items" : {
                        "format" : "binary",
                        "type" : "string"
                      },
                      "type" : "array"
                    }
                  },
                  "type" : "object"
                } ]
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/TaskCommentResponseDto"
                }
              }
            },
            "description" : "Created",
            "headers" : {
              "Location" : {
                "description" : "URL of the created resource.",
                "schema" : {
                  "format" : "uri-reference",
                  "type" : "string"
                }
              }
            }
          },
          "400" : {
            "$ref" : "#/components/responses/BadRequest"
          },
          "401" : {
            "$ref" : "#/components/responses/Unauthorized"
          },
          "403" : {
            "$ref" : "#/components/responses/Forbidden"
          },
          "404" : {
            "$ref" : "#/components/responses/NotFound"
          },
          "413" : {
            "$ref" : "#/components/responses/ContentTooLarge"
          },
          "415" : {
            "$ref" : "#/components/responses/UnsupportedMediaType"
          },
          "422" : {
            "content" : {
              "application/problem+json" : {
                "examples" : {
                  "infected-file" : {
                    "summary" : "File rejected by the antivirus",
                    "value" : {
                      "status" : 422,
                      "title" : "File rejected by the antivirus",
                      "type" : "https://github.com/jubasse/spring-boot-tasks-api/blob/main/docs/problems.md#infected-file"
                    }
                  },
                  "invalid-mention" : {
                    "summary" : "Invalid mention",
                    "value" : {
                      "status" : 422,
                      "title" : "Invalid mention",
                      "type" : "https://github.com/jubasse/spring-boot-tasks-api/blob/main/docs/problems.md#invalid-mention"
                    }
                  }
                },
                "schema" : {
                  "$ref" : "#/components/schemas/Problem"
                }
              }
            },
            "description" : "Unprocessable Content: Invalid mention, or File rejected by the antivirus."
          },
          "503" : {
            "$ref" : "#/components/responses/ServiceUnavailable"
          },
          "default" : {
            "$ref" : "#/components/responses/UnexpectedError"
          }
        },
        "summary" : "Comment on a task, as JSON or as a form with up to 5 files",
        "tags" : [ "Comments" ]
      }
    },
    "/api/v1/tasks/{id}/comments/{commentId}" : {
      "delete" : {
        "description" : "**Access:** The comment's author, or the ADMIN role.",
        "operationId" : "deleteComment",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        }, {
          "in" : "path",
          "name" : "commentId",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        } ],
        "responses" : {
          "204" : {
            "description" : "No Content"
          },
          "400" : {
            "$ref" : "#/components/responses/BadRequest"
          },
          "401" : {
            "$ref" : "#/components/responses/Unauthorized"
          },
          "403" : {
            "$ref" : "#/components/responses/Forbidden"
          },
          "404" : {
            "$ref" : "#/components/responses/NotFound"
          },
          "default" : {
            "$ref" : "#/components/responses/UnexpectedError"
          }
        },
        "summary" : "Delete a comment and its files",
        "tags" : [ "Comments" ]
      },
      "get" : {
        "operationId" : "getComment",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        }, {
          "in" : "path",
          "name" : "commentId",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/TaskCommentResponseDto"
                }
              }
            },
            "description" : "OK"
          },
          "400" : {
            "$ref" : "#/components/responses/BadRequest"
          },
          "401" : {
            "$ref" : "#/components/responses/Unauthorized"
          },
          "403" : {
            "$ref" : "#/components/responses/Forbidden"
          },
          "404" : {
            "$ref" : "#/components/responses/NotFound"
          },
          "default" : {
            "$ref" : "#/components/responses/UnexpectedError"
          }
        },
        "summary" : "Get a comment",
        "tags" : [ "Comments" ]
      },
      "patch" : {
        "description" : "**Access:** Only the comment's author.",
        "operationId" : "editComment",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        }, {
          "in" : "path",
          "name" : "commentId",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/TaskCommentDto"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/TaskCommentResponseDto"
                }
              }
            },
            "description" : "OK"
          },
          "400" : {
            "$ref" : "#/components/responses/BadRequest"
          },
          "401" : {
            "$ref" : "#/components/responses/Unauthorized"
          },
          "403" : {
            "$ref" : "#/components/responses/Forbidden"
          },
          "404" : {
            "$ref" : "#/components/responses/NotFound"
          },
          "422" : {
            "content" : {
              "application/problem+json" : {
                "examples" : {
                  "invalid-mention" : {
                    "summary" : "Invalid mention",
                    "value" : {
                      "status" : 422,
                      "title" : "Invalid mention",
                      "type" : "https://github.com/jubasse/spring-boot-tasks-api/blob/main/docs/problems.md#invalid-mention"
                    }
                  }
                },
                "schema" : {
                  "$ref" : "#/components/schemas/Problem"
                }
              }
            },
            "description" : "Unprocessable Content: Invalid mention."
          },
          "default" : {
            "$ref" : "#/components/responses/UnexpectedError"
          }
        },
        "summary" : "Edit a comment",
        "tags" : [ "Comments" ]
      }
    },
...
  },
  "security" : [ {
    "bearer" : [ ]
  } ],
  "servers" : [ {
    "url" : "/"
  } ],
...
}